the view is removed and the styles are passed to the TouchableOpacity. only padding is unified. the data type is changed. the Props interface is removed. const result is removed. switch is modified. the name of the Button component is changed to CustomButton. the year style is removed and the author name is changed to subtitle. the color and fontSize of the genre style is removed
