+++ wbooks/src/app/components/itemDetail/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Text, View, Image } from 'react-native';
-import Button from '@components/button';
+import CustomButton from '@components/customButton';
 
-import { BUTTONTHEME } from '../button/utils';
+import { BUTTONTHEME } from '../customButton/utils';
 
 import styles from './styles';
 
 const imageDefault = 'https://wtwp.com/wp-content/uploads/2015/06/placeholder-image.png';
 
-export interface Props {
+interface Props {
   title: string;
   author: string;
   imageUrl: string | null;
@@ -23,13 +23,13 @@ function ItemDetail({ title, author, imageUrl, year, genre }: Props) {
         <Image style={styles.image} source={{ uri: imageUrl ? imageUrl : imageDefault }} />
         <View style={styles.containerInfo}>
           <Text style={styles.title}>{title}</Text>
-          <Text style={styles.author}>{author}</Text>
-          <Text style={styles.year}>{year}</Text>
-          <Text style={styles.genre}>{genre}</Text>
+          <Text style={styles.subTitle}>{author}</Text>
+          <Text style={styles.subTitle}>{year}</Text>
+          <Text style={[styles.subTitle, styles.genre]}>{genre}</Text>
         </View>
       </View>
-      <Button text={'add to wishlist'} stylesButton={BUTTONTHEME.OUTLINE} />
-      <Button text={'rent'} stylesButton={BUTTONTHEME.GRADIENTS} />
+      <CustomButton text={'add to wishlist'} stylesButton={BUTTONTHEME.OUTLINE} />
+      <CustomButton text={'rent'} stylesButton={BUTTONTHEME.GRADIENTS} />
     </View>
   );
 }
